Combine all of the selfie package into one file.
